MOD: wind retrieval
Changing the philosophy of named arguments in solve_wind_field. Before I
had used every possible argument as a named agrument, but now I only use
named arguments for required arguments, and kwargs for arguments that
may or may not be specified and/or used. This was done because there was
just far too many named arguments in solve_wind_field and it looked far
too messy.