MOD: hydrometeor fall speed
Modifying the fall speed calculation. The way it was originally written
was the user provided a grid that best represented the entire domain,
called the network grid, and got back the corresponding fall velocities.
However, this is not very desirable if the network grid is from a single
radar (e.g. CSAPR) since there will not be a fall speed estimation
available where the radar has no observations, e.g. the cone of silence.

The modifcation is to provide all grids and use each grid's data to
estimate fall speed, and add this as a field to the grid object.

A possible future modification would be to add the capability to assume
one radar's observations are more reliable than the others (the original
intent of using the network) and use that radar's observations wherever
possible...