MOV: moving cost and gradient functions
Moving the wind cost and Jacobian (gradient) functions from cga.py to
winds.py. This move makes cga.py basically obsolete.

I was trying to think of a more clever way to design the cost and
Jacobian, but in order for these to be inputs to the SciPy minimize
module these two need to be callable functions which take only the
analysis vector (x) and positional arguments (*args) as inputs. So
right now I am not certain if I can create a nice class for these
two functions...