MOD: major modifications
This commit includes some major modifications, listed below:

1. The cost function (_wind_cost) and Jacobian (_wind_grad) are now
found in winds.py, making cga.py basically obsolete.
2. The impermeability condition is now available as a strong constraint
rather than a weak constraint. My goal is to make it possible for either
types of constraints, but for now it is only as a strong constraint.
3. The ability to use a standard atmospheric profile of temperature,
pressure, and density, allowing the user to not input a sounding
dataset. A little more work needs to be done on this though...
4. Smarter routines that recognize when certain variables are not
required, e.g., echo base and top heights are only required when the
continuity cost is set to 'iterative'.
5. The background constraint, continuity, and smoothness constraints can
be ignored with intuitive parameter settings, not just by setting their
weights to zero. E.g., smooth_cost = None, continuity_cost = None, and
background = None.