ENH: several enhancements
I am committing several changes here, the major of which fixes how the
impermeability condition is handled. What I was doing before was not
working properly, so it had to be fixed. Another big enhancement is in
the new options available to the user for how the continuity equation
can be used. There are currently four options: 'Potvin', 'Protat',
'bottom-up', and 'top-down'.