MOD: a few modifications
A few changes including the documentation of solve_wind_field(). I am
also adding the parameter maxiter_first_pass which allows the user to
input a different maximum number of iterations for the first pass
minimization (if applicable). Generally the first pass does not require
as many iterations as the "full" retrieval.